backend/pyproject.toml
[tool.poetry]
name = "landgriffon-mktp"
version = "0.1.0"
description = "landgriffon marketplace subscription app"
authors = ["Miguel Mendozo <miguel.mendoza@vizzuality.com>"]
license = "GNU"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.30"
google-cloud-pubsub = "^2.21.3"
pydantic = "^2.7.4"
python-dotenv = "^1.0.1"
logging = "^0.4.9.6"
alembic = "^1.13.1"
google-api-python-client = "^2.133.0"
google-auth = "^2.30.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
